requirements.txt
discord.py
google-generativeai
python-dotenv
Pillow
elevenlabs
Unidecode
Flask
